requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==9.5.0
spacy==3.5.1
torch==2.0.1
transformers==4.31.0
torchvision==0.16.0
